requirements.txt
absl-py==1.0.0
amqp==5.0.9
asgiref==3.4.1
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0
djangorestframework==3.13.1
iniconfig==1.1.1
kombu==5.2.3
numpy==1.22.0
ortools==9.2.9972
packaging==21.3
pika==1.2.0
pluggy==1.0.0
prompt-toolkit==3.0.24
protobuf==3.19.1
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytz==2021.3
redis==4.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
